Switch WSGIApplication example to Pyramid
The SDK provides a first-class integration for Flask, so using Flask as
the example for the generic WSGIApplication integration for the SDK
might be confusing to some users. This change rewrites the example to
use a "hello world" application example for Pyramid, one of the most
popular web frameworks for Python to remove the potential confusion.