Make logging from Flask more explicit
We got some user feedback that their logs weren't being sent to
Application Insights from the Flask integration. This was due to a
confusion between the Python rooter logger `logging.debug(...)` and the
Flask loggers `app.logger.debug(...)`. The Flask integration only sends
the latter logs to Application Insights since mutating the root logger
from a third party package would be surprising behavior.

This change makes it explicit in the documentation how the logger has to
be used so that the messages to end up in Application Insights.