Ensure CI and PyPI Python versions match
The classifiers list in `setup.py` mentions support for Python 3.4 and
yet the CI on Travis doesn't run the tests on that version which means
that we can't be certain that the package works correctly on the
declared version.

To fix this discrepancy, this change makes the Python versions used
during CI match up with the Python versions declared in the classifiers
list.