Add option to make the telemetry queue persistent
Currently the Application Insights client SDK already has some great
resiliency built-in, e.g. retrying to send telemetry if there are
intermittent connectivity issues or service interruptions. However,
there still is a risk to lose telemetry if the app crashes before
sending telemetry since the telemetry queue is in-memory only.

This change adds an option to persist the queue to disk via the
[persist-queue](https://pypi.org/project/persist-queue/) module which
will protect customers against potential data loss, at the cost of a
slight loss in performance.