Fix bug when NullSender is used with Async Queue
Currently the NullSender crashes when used in conjunction with
AsynchronousQueue since the AsynchronousQueue relies on some methods
that are not part of the SenderBase interface.

This limitation leads to unexpected behavior when for example using the
NullSender to avoid sending telemetry in an application when the
instrumentation key isn't configured in the environment.

This change makes the NullSender work with the AsynchronousQueue so that
the class can be used with both queues included in the SDK.